Deprecate the `simple` species type
This is done in favor of more explicit species types such as e.g.
`Electron`.
