+++ packages/schema/src/species/particles/particle.ts
@@ -1,0 +1,8 @@
+// SPDX-FileCopyrightText: LXCat team
+//
+// SPDX-License-Identifier: Apache-2.0
+
+export type Particle = {
+  composition: string;
+  charge: number;
+};
