+++ packages/schema/src/species/particles/serialize.ts
@@ -2,18 +2,19 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-import { SimpleParticle } from "../composition/simple/particle.js";
-import { serializeSimpleParticle } from "../composition/simple/serialize.js";
-import { compositionSummary } from "../composition/universal.js";
+import { parseCharge, parseChargeLatex } from "../common.js";
 import { StateSummary } from "../summary.js";
-import { AnyParticle } from "./any-particle.js";
+import { Particle } from "./particle.js";
 
-export const serializeAnyParticle = (particle: AnyParticle): StateSummary => {
-  if (typeof particle.composition === "string") {
-    return serializeSimpleParticle(particle as SimpleParticle);
-  }
-
-  const composition = compositionSummary(particle.composition, particle.charge);
-
-  return { composition, ...composition };
+export const serializeSimpleParticle = (
+  state: Particle,
+): StateSummary => {
+  const composition = {
+    summary: `${state.composition}${parseCharge(state.charge)}`,
+    latex: `\\mathrm{${state.composition}}${parseChargeLatex(state.charge)}`,
+  };
+  return ({
+    composition,
+    ...composition,
+  });
 };
